+++ docs/CHANGELOG.md
@@ -16,6 +16,7 @@
 - frontend: introduced fully featured markdown-support with extensions
 - frontend: introduced print/pdf-support
 - frontend: make map-components generic
+- frontend: make imageobject-components generic
 
 ### new features
 - moved TextEditorComponent, FormUtils and GeoParserDeterminer to commons
@@ -23,6 +24,7 @@
 - frontend: introduced fully featured markdown-support with extensions
 - frontend: introduced print/pdf-support
 - frontend: make map-components generic
+- frontend: make imageobject-components generic
 
 ### improvements
 - moved layout-utils, noop-action, modal-functionality to commons